requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
llama-cpp-python>=0.2.50 # Or newer. Check compatibility!
python-dotenv>=1.0.0
sse-starlette>=1.6.5     # For easy Server-Sent Events streaming

# Optional for Nice-to-Haves
# slowapi>=0.1.8         # For rate limiting
# typer>=0.9.0           # For CLI tool